pyproject.toml
[project]
name = "bida-ml-starter"
version = "0.1.0"
description = "Pistor BIDA ML Starter Template — Forecasting, Klassifikation, Regression"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    # Data
    "snowflake-connector-python[pandas]>=3.12",
    "snowflake-sqlalchemy>=1.6",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "openpyxl>=3.1.0",
    "pyarrow>=15.0.0",

    # ML Core
    "scikit-learn>=1.5.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "optuna>=3.6.0",
    "imbalanced-learn>=0.12.0",

    # Time Series
    "statsmodels>=0.14.0",
    "neuralforecast>=1.7.0",
    "statsforecast>=2.0.0",
    "mlforecast>=0.13.0",

    # Explainability
    "shap>=0.45.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
    "scipy>=1.14.0",
    "holidays>=0.50",
    "joblib>=1.4.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.0.0",
    "jupyterlab>=4.2.0",
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "nbstripout>=0.7.0",
]
